modal_endpoint_app/requirements.txt
modal==1.1.3
fastapi[standard]==0.117.1
uvicorn==0.37.0
pydantic==2.10.3
chromadb==1.1.0
numpy==2.2.6
sentence-transformers==5.1.1
python-dotenv==1.1.1
PyMuPDF==1.26.3
openai==2.7.1
langchain-openai==0.3.35
langchain==0.3.27